make all resources under project supported
