+++ codegen-renderers/src/main/kotlin/io/vrap/codegen/languages/typescript/client/AbstractRequestBuilder.kt
@@ -6,6 +6,7 @@ import io.vrap.codegen.languages.php.extensions.EObjectTypeExtensions
 import io.vrap.codegen.languages.php.extensions.resource
 import io.vrap.codegen.languages.php.extensions.toResourceName
 import io.vrap.codegen.languages.typescript.client.files_producers.middleware
+import io.vrap.codegen.languages.typescript.model.TypeScriptBaseTypes
 import io.vrap.codegen.languages.typescript.model.simpleTSName
 import io.vrap.codegen.languages.typescript.tsMediaType
 import io.vrap.codegen.languages.typescript.tsRemoveRegexp
@@ -14,7 +15,9 @@ import io.vrap.codegen.languages.typescript.tsRequestName
 import io.vrap.rmf.codegen.types.VrapObjectType
 import io.vrap.rmf.codegen.types.VrapTypeProvider
 import io.vrap.rmf.raml.model.modules.Api
+import io.vrap.rmf.raml.model.resources.Method
 import io.vrap.rmf.raml.model.resources.Resource
+import io.vrap.rmf.raml.model.resources.ResourceContainer
 import java.nio.file.Path
 import java.nio.file.Paths
 
@@ -23,96 +26,7 @@ abstract class AbstractRequestBuilder constructor(
         override val vrapTypeProvider: VrapTypeProvider
 ) : EObjectTypeExtensions {
 
-
-    protected fun Resource.constructor(): String {
-
-        val pathArgs = if (this.fullUri.variables.isEmpty())
-            ""
-        else
-            this
-                    .fullUri
-                    .variables
-                    .map { "      $it: string" }
-                    .joinToString(
-                            separator = ",\n",
-                            prefix = "pathArgs: {\n",
-                            postfix = "\n }")
-
-        return """|
-                    |  constructor(
-                    |    protected readonly args: {
-                    |      <$pathArgs>,
-                    |      middlewares: Middleware[];
-                    |    }
-                    |  ) {}
-                    """.trimMargin()
-    }
-
-    protected fun Resource.methods(): String {
-
-        return this.methods
-                .map {
-                    val methodArgs = if (it.queryParameters.isNullOrEmpty() && it.bodies.isNullOrEmpty())
-                        ""
-                    else {
-                        var queryParamsArg = ""
-                        var bodies = ""
-                        if (!it.queryParameters.isNullOrEmpty()) {
-                            val allQueryParamsOptional = it.queryParameters.map { !it.required }.reduce(Boolean::and)
-                            queryParamsArg =
-                                    """|queryArgs${if (allQueryParamsOptional) "?" else ""}: {
-                            |   <${it.queryParameters.map { "${it.name.tsRemoveRegexp()}${if (it.required) "" else "?"}: ${it.type.toVrapType().simpleTSName()}" }.joinToString(separator = "\n")}>
-                            |},""".trimMargin()
-                        }
-                        if (!it.bodies.isNullOrEmpty()) {
-                            bodies = "payload: ${it.bodies.map {
-                                it.type.toVrapType().simpleTSName()
-                            }.joinToString(separator = " | ")}"
-                        }
-
-                        val argsAreOptional =
-                                it.bodies.isNullOrEmpty()
-                                        && (
-                                        it.queryParameters.isNullOrEmpty() ||
-                                                it.queryParameters.map { !it.required }.reduce(Boolean::and)
-                                        )
-
-                        """ |
-                            |methodArgs${if (argsAreOptional) "?" else ""}:{
-                            |   <$queryParamsArg>
-                            |   <$bodies>
-                            |}
-                        """.trimMargin()
-                    }
-                    val methodReturn = "ApiRequest\\<${it.bodies.map { it.type.toVrapType().simpleTSName() }.joinToString(separator = " | ").ifEmpty { "void" }}, ${it.returnType().toVrapType().simpleTSName()}, ${it.tsRequestName()}\\>"
-
-                    val bodyLiteral = """|{
-                        |   baseURL: '${api.baseUri.template}',
-                        |   method: '${it.methodName.toUpperCase()}',
-                        |   uriTemplate: '${it.resource().fullUri.template}',
-                        |   pathVariables: this.args.pathArgs,
-                        |   headers: {
-                        |       <${if(it.tsMediaType().isNotEmpty()) "${it.tsMediaType()}," else ""}>
-                        |   },
-                        |   <${if(it.queryParameters.isNullOrEmpty()) "" else "queryParams: (methodArgs || {} as any).queryArgs,"}>
-                        |   <${if(it.bodies.isNullOrEmpty()) "" else "payload: (methodArgs || {} as any).payload,"}>
-                        |}
-                    """.trimMargin()
-
-
-                    """|
-                    |${it.methodName}(<$methodArgs>): $methodReturn {
-                    |   return new $methodReturn(
-                    |       <$bodyLiteral>,
-                    |       this.args.middlewares
-                    |   )
-                    |}
-                    |
-                 """.trimMargin()
-                }.joinToString(separator = "")
-    }
-
-    protected fun Resource.subResources(): String {
+    protected fun ResourceContainer.subResources(): String {
         return this.resources
                 .map {
 
@@ -128,7 +42,7 @@ abstract class AbstractRequestBuilder constructor(
                     |   return new ${it.toRequestBuilderName()}(
                     |         {
                     |            pathArgs: {
-                    |               ...this.args.pathArgs,
+                    |               <${if(hasPathArgs()) "...this.args.pathArgs," else ""}>
                     |               <${if (it.relativeUri.variables.isNotEmpty()) "...childPathArgs" else ""}>
                     |            },
                     |            middlewares: this.args.middlewares
@@ -140,54 +54,10 @@ abstract class AbstractRequestBuilder constructor(
                 }.joinToString(separator = "")
     }
 
-    fun Resource.imports(moduleName: String): String {
-        return this.methods
-                .map {
-                    val requestModuleName = it.tsRequestModuleName((it.toVrapType() as VrapObjectType).`package`)
-                    val relativePath = relativizePaths(moduleName, requestModuleName)
-                    "import { ${it.tsRequestName()} } from '$relativePath'"
-                }.plus(
-                        this.resources
-                                .map {
-                                    val relativePath = relativizePaths(moduleName, it.tsRequestModuleName((this.toVrapType() as VrapObjectType).`package`))
-                                    "import { ${it.toRequestBuilderName()} } from '$relativePath'"
-                                }
-                ).plus(
-                        this.methods
-                                .flatMap { it.bodies }
-                                .filter { it.type != null }
-                                .map { it.type.toVrapType() }
-                                .filter { it is VrapObjectType }
-                                .map { it as VrapObjectType }
-                                .map {
-                                    val relativePath = relativizePaths(moduleName, it.`package`)
-                                    "import { ${it.simpleTSName()} } from '$relativePath'"
-                                }
-                )
-                .plus(
-                        this.methods
-                                .map { it.returnType().toVrapType() }
-                                .filter { it is VrapObjectType }
-                                .map { it as VrapObjectType }
-                                .map {
-                                    val relativePath = relativizePaths(moduleName, it.`package`)
-                                    "import { ${it.simpleTSName()} } from '$relativePath'"
-                                }
-                )
-                .plus(
-                        "import { ${middleware.simpleClassName} } from '${relativizePaths(moduleName, middleware.`package`)}'"
-                )
-                .distinct()
-                .joinToString(separator = "\n")
-    }
-
-    protected fun relativizePaths(currentModule: String, targetModule: String): String {
-        val currentRelative: Path = Paths.get(currentModule.replace(".", "/"))
-        val targetRelative: Path = Paths.get(targetModule.replace(".", "/"))
-        return "./" + currentRelative.relativize(targetRelative).toString().replaceFirst("../", "")
-    }
-
     protected fun Resource.toRequestBuilderName(): String = "${this.toResourceName()}RequestBuilder"
+
+
+    abstract fun hasPathArgs(): Boolean
 }
