+++ codegen-renderers/src/main/kotlin/io/vrap/codegen/languages/typescript/client/files_producers/ApiRootFileProducer.kt
@@ -3,13 +3,17 @@ package io.vrap.codegen.languages.typescript.client.files_producers
 import com.google.inject.Inject
 import com.google.inject.name.Named
 import io.vrap.codegen.languages.typescript.client.AbstractRequestBuilder
+import io.vrap.codegen.languages.typescript.model.TypeScriptBaseTypes
+import io.vrap.codegen.languages.typescript.model.simpleTSName
+import io.vrap.codegen.languages.typescript.relativizePaths
+import io.vrap.codegen.languages.typescript.tsRequestModuleName
 import io.vrap.rmf.codegen.di.VrapConstants
 import io.vrap.rmf.codegen.io.TemplateFile
 import io.vrap.rmf.codegen.rendring.FileProducer
 import io.vrap.rmf.codegen.rendring.utils.keepIndentation
+import io.vrap.rmf.codegen.types.VrapObjectType
 import io.vrap.rmf.codegen.types.VrapTypeProvider
 import io.vrap.rmf.raml.model.modules.Api
-import io.vrap.rmf.raml.model.resources.Resource
 
 class ApiRootFileProducer @Inject constructor(
         @Named(VrapConstants.CLIENT_PACKAGE_NAME) val client_package: String,
@@ -18,32 +22,51 @@ class ApiRootFileProducer @Inject constructor(
 ) : FileProducer, AbstractRequestBuilder(api, vrapTypeProvider) {
 
 
-    val rootResources = api.resources
-
     override fun produceFiles(): List<TemplateFile> {
 
-        return listOf(produceApiRoot(rootResources[0]))
+        return listOf(produceApiRoot(api))
     }
 
-    fun produceApiRoot(type: Resource): TemplateFile {
-        val modeuleName = "$client_package.ApiRoot"
+    fun produceApiRoot(type: Api): TemplateFile {
+        val moduleName = "$client_package.ApiRoot"
         return TemplateFile(
-                relativePath = modeuleName.replace(".", "/") + ".ts",
+                relativePath = moduleName.replace(".", "/") + ".ts",
                 content = """|
-                |${type.imports(modeuleName)}
-                |import { ApiRequest } from '${relativizePaths(modeuleName,"base.requests-utils")}'
+                |${type.imports(moduleName)}
+                |import { ApiRequest } from '${relativizePaths(moduleName, "base.requests-utils")}'
                 |
                 |export class ApiRoot {
                 |
-                |    <${type.constructor()}>
-                |    <${type.subResources()}>
-                |    <${type.methods()}>
+                |  constructor(
+                |    protected readonly args: {
+                |      middlewares: Middleware[];
+                |    }
+                |  ) {}
+                |
+                |  <${type.subResources()}>
                 |
                 |}
                 |
             """.trimMargin().keepIndentation()
 
         )
-
     }
+
+    fun Api.imports(moduleName: String): String {
+        return this.resources
+                .map {
+                    val relativePath = relativizePaths(moduleName, it.tsRequestModuleName((it.toVrapType() as VrapObjectType).`package`))
+                    "import { ${it.toRequestBuilderName()} } from '$relativePath'"
+                }
+                .plus(
+                        "import { ${middleware.simpleClassName} } from '${relativizePaths(moduleName, middleware.`package`)}'"
+                )
+                .plus(
+                        "import { ${TypeScriptBaseTypes.file.simpleTSName()} } from '${relativizePaths(moduleName, (TypeScriptBaseTypes.file as VrapObjectType).`package`)}'"
+                )
+                .distinct()
+                .joinToString(separator = "\n")
+    }
+
+    override fun hasPathArgs(): Boolean = false
 }
