+++ codegen-renderers/src/main/kotlin/io/vrap/codegen/languages/typescript/client/files_producers/ConstEntities.kt
@@ -6,4 +6,3 @@ import io.vrap.rmf.codegen.types.VrapObjectType
 val commonRequest = VrapObjectType("base.common-types","CommonRequest")
 val apiRequest = VrapObjectType("base.requests-utils","ApiRequest")
 val middleware = VrapObjectType("base.common-types","Middleware")
-
