+++ codegen-renderers/src/main/kotlin/io/vrap/codegen/languages/typescript/model/TypeScriptObjectTypeExtensions.kt
@@ -71,3 +71,5 @@ private fun toVrapType(vrapType: VrapType): VrapType? {
 
     }
 }
+
+
