+++ codegen-renderers/src/main/kotlin/io/vrap/codegen/languages/typescript/model/TypeScriptVrapExtensions.kt
@@ -12,3 +12,4 @@ fun VrapType.simpleTSName():String{
         is VrapNilType -> this.name
     }
 }
+
